Introduce EmptyHypha
